+++ README.md
@@ -3,14 +3,14 @@
 ## step1
 
 ### UI
-- [ ] 영화를 리스트에 보여준다.
-  - [ ] 영화 예매 버튼을 누르면 상세 정보 화면으로 이동한다.
-- [ ] 영화의 상세 정보를 보여준다.
-  - [ ] 예약할 인원을 선택할 수 있다.
-  - [ ] 예매 완료 버튼을 누르면 예약 완료 화면으로 이동한다.
-  - [ ] 뒤로가기 버튼을 누르면 영화 목록 화면으로 이동한다.
-- [ ] 최종 예약 정보를 보여준다.
-  - [ ] 뒤로가기 버튼을 누르면 영화 목록 화면으로 이동한다.
+- [x] 영화를 리스트에 보여준다.
+  - [x] 영화 예매 버튼을 누르면 상세 정보 화면으로 이동한다.
+- [x] 영화의 상세 정보를 보여준다.
+  - [x] 예약할 인원을 선택할 수 있다.
+  - [x] 예매 완료 버튼을 누르면 예약 완료 화면으로 이동한다.
+  - [x] 뒤로가기 버튼을 누르면 영화 목록 화면으로 이동한다.
+- [x] 최종 예약 정보를 보여준다.
+  - [x] 뒤로가기 버튼을 누르면 영화 목록 화면으로 이동한다.
 
 ### Domain
-- [ ] 구매할 티켓에 대한 가격을 계산한다.
+- [x] 구매할 티켓에 대한 가격을 계산한다.
