+++ app/src/main/java/woowacourse/movie/view/fragments/MovieListFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,9 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
-import woowacourse.movie.adapter.MovieListAdapter
+import woowacourse.movie.R
+import woowacourse.movie.commit
 import woowacourse.movie.data.Movie
 import woowacourse.movie.databinding.FragmentMovieListBinding
+import woowacourse.movie.view.adapter.MovieListAdapter
 
 class MovieListFragment : Fragment() {
     private var _binding: FragmentMovieListBinding? = null
