+++ app/src/main/java/woowacourse/movie/view/fragments/TicketingResultFragment.kt
@@ -1,20 +1,21 @@
-package woowacourse.movie
+package woowacourse.movie.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import woowacourse.movie.MovieListFragment.Companion.MOVIE_KEY
-import woowacourse.movie.TicketingFragment.Companion.TICKET_COUNT_KEY
+import woowacourse.movie.R
 import woowacourse.movie.data.Movie
 import woowacourse.movie.databinding.FragmentTicketingResultBinding
+import woowacourse.movie.domain.Ticket
+import woowacourse.movie.getSerializable
+import woowacourse.movie.view.fragments.MovieListFragment.Companion.MOVIE_KEY
+import woowacourse.movie.view.fragments.TicketingFragment.Companion.TICKET_KEY
 
 class TicketingResultFragment : Fragment() {
     private var _binding: FragmentTicketingResultBinding? = null
     private val binding get() = _binding!!
-
-    private val ticketCount: Int by lazy { arguments?.getInt(TICKET_COUNT_KEY) ?: 0 }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -32,9 +33,16 @@ class TicketingResultFragment : Fragment() {
                 tvTitle.text = it.title
                 tvDate.text = getString(R.string.movie_release_date, it.formattedDate)
             }
-            tvRegularCount.text = getString(R.string.regular_count, ticketCount)
-            tvPayResult.text =
-                getString(R.string.movie_pay_result, ticketCount * 13000, "현장 결제")
+
+            getSerializable<Ticket>(TICKET_KEY)?.let {
+                tvRegularCount.text = getString(R.string.regular_count, it.count)
+                tvPayResult.text =
+                    getString(
+                        R.string.movie_pay_result,
+                        it.calculateTotalPrice(),
+                        getString(R.string.on_site_payment)
+                    )
+            }
         }
     }
