+++ app/src/main/res/layout/activity_main.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".view.activities.MainActivity">
 
     <fragment
         android:id="@+id/fragment_movie"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:name="woowacourse.movie.MovieListFragment" />
+        android:name="woowacourse.movie.view.fragments.MovieListFragment" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
