+++ app/src/main/res/layout/fragment_ticketing.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TicketingFragment">
+    tools:context=".view.fragments.TicketingFragment">
 
     <ImageView
         android:id="@+id/iv_poster"
