+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,5 @@
     <string name="ticketing_cancel_info">영화 상영 시작 시간 15분 전까지 취소가 가능합니다.</string>
     <string name="regular_count">일반 %d명</string>
     <string name="movie_pay_result">%,d원 (%s)</string>
+    <string name="on_site_payment">현장 결제</string>
 </resources>
